Create wxmsig.sol
If you want to create a smart contract A with multi-sig functions, you should write "contract A is multiowned{}". If you want to add time limits in your contract A, you can write "contract A is daylimit, multiowned{}".  Given a function B needs to be multi-sighed in the contract A, you can write as the following,

contract A is daylimit, multi owned{
      
      function B ( ) onlymanyowners(sha3(msg.data)) external {
    
           }
}